+++ src/main/java/org/folio/ed/controller/LookupAsrRequestsController.java
@@ -1,6 +1,5 @@
 package org.folio.ed.controller;
 
-import jakarta.validation.constraints.NotNull;
 import org.folio.ed.service.RemoteStorageService;
 import org.folio.ed.domain.dto.AsrRequests;
 import org.folio.ed.rest.resource.LookupAsrRequestsApi;
@@ -9,11 +8,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
@@ -26,10 +22,8 @@ public class LookupAsrRequestsController implements LookupAsrRequestsApi {
   private final RemoteStorageService remoteStorageService;
 
   @Override
-  public ResponseEntity<AsrRequests> getAsrRequests(
-      @ApiParam(required = true) @PathVariable("remoteStorageConfigurationId") String remoteStorageConfigurationId,
-      @NotNull @ApiParam(required = true) @RequestHeader(value = "x-okapi-token") String xOkapiToken,
-      @NotNull @ApiParam(required = true) @RequestHeader(value = "x-okapi-tenant") String xOkapiTenant) {
+  public ResponseEntity<AsrRequests> getAsrRequests(String remoteStorageConfigurationId,
+      String xOkapiToken, String xOkapiTenant) {
 
     var headers = new HttpHeaders();
     headers.setContentType(MediaType.APPLICATION_XML);