+++ src/main/java/org/folio/ed/controller/UpdateASRItemStatusBeingRetrievedController.java
@@ -1,7 +1,5 @@
 package org.folio.ed.controller;
 
-import io.swagger.annotations.ApiParam;
-import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.ed.domain.dto.UpdateAsrItem;
@@ -10,12 +8,7 @@ import org.folio.ed.service.RemoteStorageService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-
-import jakarta.validation.Valid;
 
 @Log4j2
 @Controller
@@ -26,11 +19,8 @@ public class UpdateASRItemStatusBeingRetrievedController implements UpdateASRIte
   private final RemoteStorageService remoteStorageService;
 
   @Override
-  public ResponseEntity<Void> updateAsrItemCheckIn(
-    @ApiParam(required = true) @PathVariable("remoteStorageConfigurationId") String remoteStorageConfigurationId,
-    @NotNull @ApiParam(required = true) @RequestHeader(value = "x-okapi-token") String xOkapiToken,
-    @NotNull @ApiParam(required = true) @RequestHeader(value = "x-okapi-tenant") String xOkapiTenant,
-    @ApiParam(required = true) @Valid @RequestBody UpdateAsrItem updateAsrItem) {
+  public ResponseEntity<Void> updateAsrItemCheckIn(String remoteStorageConfigurationId,
+    String xOkapiToken, String xOkapiTenant, UpdateAsrItem updateAsrItem) {
     remoteStorageService.checkInItemByBarcode(remoteStorageConfigurationId, updateAsrItem.getItemBarcode(), xOkapiTenant,
       xOkapiToken);
     return new ResponseEntity<>(HttpStatus.CREATED);